graphql tests refactors
