+++ playwright/tests/specs/graphql/udg_fe_internal_sources.spec.ts
@@ -1,11 +1,9 @@
 import { test, assert } from '@fixtures';
-import { apis_page } from '../../../lib/pom/Apis_page';
-import { graphql_page } from '../../../lib/pom/Graphql_page';
-
+import { Locator } from '@playwright/test';
 import { Dashboard_connection } from '@api_connections/Dashboard_connection';
 import { newAPIdefinitionWithDefaults } from '@lib/utils/API_object_designer';
 
-test('UDG with internal REST and GQL datasources', async ({ createUserAndLogin, main_page }) => {
+test('UDG with internal REST and GQL datasources', async ({ createUserAndLogin, main_page, apis_page, graphql_page, page }) => {
 
     const udgDetails = {
         apiName: "UDG-InternalSources",
@@ -51,7 +49,7 @@ test('UDG with internal REST and GQL datasources', async ({ createUserAndLogin, 
     };
 
     const gqlApi = {
-        "name": "gqlApi",        
+        "name": "gqlApi",
         "graphql": {
             "enabled": true,
             "execution_mode": "proxyOnly",
@@ -64,66 +62,66 @@ test('UDG with internal REST and GQL datasources', async ({ createUserAndLogin, 
     };
 
     const schemaFileRelativePath = "../../test/specs/graphql/udg-schema.gql";
-    let $apiTableElement;
-    
+    let $apiTableElement: Locator;
+
     let refreshCounter = 0;
-    
+
     const dashboard_connection = new Dashboard_connection();
 
-    xawait test.step('Prerequisites: creating APIs for internal datasources via dashboard API', async () => {
+    await test.step('Prerequisites: creating APIs for internal datasources via dashboard API', async () => {
         [restApi, gqlApi].forEach(api => {
             let body = newAPIdefinitionWithDefaults(api);
             await dashboard_connection.createAPI(body, createUserAndLogin.userSecret);
         })
     });
 
-    xawait test.step('User should be able to create a UDG API with internal REST and GraphQL datasources', async () => {
+    await test.step('User should be able to create a UDG API with internal REST and GraphQL datasources', async () => {
         await main_page.openAPIs();
-        while(!apis_page.ADD_NEW_API_BUTTON.isExisting() && refreshCounter < 5){
-            await page.reload();
-            browser.pause(2000);
-            refreshCounter++;
-        }
-       await apis_page.ADD_NEW_API_BUTTON.click();
-       await apis_page.API_NAME_INPUT.fill(udgDetails.apiName);
-       await apis_page.API_TYPE_UDG_BUTTON.click();
-       await apis_page.CONFIGURE_API_BUTTON.click();
-       await graphql_page.GRAPHQL_SCHEMA_TAB_BUTTON.click();
-        graphql_page.uploadSchemaFile(schemaFileRelativePath);
-       await apis_page.SAVE_BUTTON.click();
+        // while(!apis_page.ADD_NEW_API_BUTTON.isExisting() && refreshCounter < 5){
+        //     await page.reload();
+        //     // browser.pause(2000);
+        //     refreshCounter++;
+        // }
+        await apis_page.ADD_NEW_API_BUTTON.click();
+        await apis_page.API_NAME_INPUT.fill(udgDetails.apiName);
+        await apis_page.API_TYPE_UDG_BUTTON.click();
+        await apis_page.CONFIGURE_API_BUTTON.click();
+        await graphql_page.GRAPHQL_SCHEMA_TAB_BUTTON.click();
+        await graphql_page.uploadSchemaFile(schemaFileRelativePath);
+        await apis_page.SAVE_BUTTON.click();
         await main_page.openAPIs();
-        $apiTableElement = await this.page.locator(`span=${udgDetails.apiName}`);
-      await $apiTableElement.click();
-       await graphql_page.GRAPHQL_SCHEMA_TAB_BUTTON.click();
+        $apiTableElement = page.locator(`span:text-is('${udgDetails.apiName}')`);
+        await $apiTableElement.click();
+        await graphql_page.GRAPHQL_SCHEMA_TAB_BUTTON.click();
         //Define internal REST data source for a Query type field
-        graphql_page.getUDG_OPEN_FIELD_OPTIONS_BUTTON("Query",await udgDetails.restQuery).click();
+        graphql_page.getUDG_OPEN_FIELD_OPTIONS_BUTTON("Query", await udgDetails.restQuery).click();
         graphql_page.UDG_EXPAND_REST_TYK_ACCORDION.expand();
-       await graphql_page.UDG_SELECT_REST_API_OPEN_COMBOBOX.click();
+        await graphql_page.UDG_SELECT_REST_API_OPEN_COMBOBOX.click();
         graphql_page.UDG_COMBOBOX_DROPDOWN.selectComboboxOption(restApi.name);
-       await graphql_page.UDG_DATA_SOURCE_NAME_INPUT.fill(udgDetails.restSource);
-       await graphql_page.UDG_DATA_SOURCE_ENDPOINT_INPUT.fill(udgDetails.restApiEndpoint);
-       await graphql_page.UDG_DATA_SOURCE_METHOD.selectOption("GET");
-       await graphql_page.UDG_DATA_SOURCE_SAVEANDUPDATE_BUTTON.click();
-       await apis_page.CONFIRM_BUTTON.click();
+        await graphql_page.UDG_DATA_SOURCE_NAME_INPUT.fill(udgDetails.restSource);
+        await graphql_page.UDG_DATA_SOURCE_ENDPOINT_INPUT.fill(udgDetails.restApiEndpoint);
+        await graphql_page.UDG_DATA_SOURCE_METHOD.selectOption("GET");
+        await graphql_page.UDG_DATA_SOURCE_SAVEANDUPDATE_BUTTON.click();
+        await apis_page.CONFIRM_BUTTON.click();
         await assert(graphql_page.getUDG_FIELD_DATA_SOURCE_LABEL_NAME("Query", udgDetails.restQuery, udgDetails.restSource)).toBeVisible();
         await assert(graphql_page.getUDG_FIELD_DATA_SOURCE_LABEL_TYPE("Query", udgDetails.restQuery, "REST")).toBeVisible();
         //Define internal GQL data source for a Query type field
-        graphql_page.getUDG_OPEN_FIELD_OPTIONS_BUTTON("Query",await udgDetails.gqlQuery).click();
+        graphql_page.getUDG_OPEN_FIELD_OPTIONS_BUTTON("Query", await udgDetails.gqlQuery).click();
         graphql_page.UDG_EXPAND_GRAPHQL_TYK_ACCORDION.expand();
-       await graphql_page.UDG_SELECT_GRAPHQL_API_OPEN_COMBOBOX.click();
+        await graphql_page.UDG_SELECT_GRAPHQL_API_OPEN_COMBOBOX.click();
         graphql_page.UDG_COMBOBOX_DROPDOWN.selectComboboxOption(gqlApi.name);
-       await graphql_page.UDG_DATA_SOURCE_NAME_INPUT.fill(udgDetails.gqlSource);
-       await graphql_page.UDG_DATA_SOURCE_SAVEANDUPDATE_BUTTON.click();
-       await apis_page.CONFIRM_BUTTON.click();
+        await graphql_page.UDG_DATA_SOURCE_NAME_INPUT.fill(udgDetails.gqlSource);
+        await graphql_page.UDG_DATA_SOURCE_SAVEANDUPDATE_BUTTON.click();
+        await apis_page.CONFIRM_BUTTON.click();
         await assert(graphql_page.getUDG_FIELD_DATA_SOURCE_LABEL_NAME("Query", udgDetails.gqlQuery, udgDetails.gqlSource)).toBeVisible();
         await assert(graphql_page.getUDG_FIELD_DATA_SOURCE_LABEL_TYPE("Query", udgDetails.gqlQuery, "GraphQL")).toBeVisible();
     });
 
-    xawait test.step('User should be able to remove an internal data source from an object', async () => {
-        graphql_page.getUDG_OPEN_FIELD_OPTIONS_BUTTON("Query",await udgDetails.restQuery).click();
-       await graphql_page.UDG_DATA_SOURCE_RESET_BUTTON.click();
-       await graphql_page.UDG_DATA_SOURCE_SAVEANDUPDATE_BUTTON.click();
-       await apis_page.CONFIRM_BUTTON.click();
+    await test.step('User should be able to remove an internal data source from an object', async () => {
+        graphql_page.getUDG_OPEN_FIELD_OPTIONS_BUTTON("Query", await udgDetails.restQuery).click();
+        await graphql_page.UDG_DATA_SOURCE_RESET_BUTTON.click();
+        await graphql_page.UDG_DATA_SOURCE_SAVEANDUPDATE_BUTTON.click();
+        await apis_page.CONFIRM_BUTTON.click();
         await assert(graphql_page.getUDG_FIELD_DATA_SOURCE_LABEL_NAME("Query", udgDetails.restQuery, udgDetails.restSource)).not.toBeVisible();
         await assert(graphql_page.getUDG_FIELD_DATA_SOURCE_LABEL_TYPE("Query", udgDetails.restQuery, "REST")).not.toBeVisible();
     });
